+++ framework/helpers/IpHelper.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace yii\helpers;
+
+/**
+ * Class IpHelper
+ *
+ * @author Dmytro Naumenko <d.naumenko.a@gmail.com>
+ */
+class IpHelper extends BaseIpHelper
+{
+
+}