+++ framework/validators/IpValidator.php
@@ -10,6 +10,7 @@ namespace yii\validators;
 use Yii;
 use yii\base\InvalidConfigException;
 use yii\helpers\Html;
+use yii\helpers\IpHelper;
 use yii\helpers\Json;
 use yii\web\JsExpression;
 
@@ -354,14 +355,14 @@ class IpValidator extends Validator
             return [$this->message, []];
         }
 
-        if ($this->getIpVersion($ip) == 6) {
+        if ($this->getIpVersion($ip) === IpHelper::IPV6) {
             if ($cidr !== null) {
-                if ($cidr > static::IPV6_ADDRESS_LENGTH || $cidr < 0) {
+                if ($cidr > IpHelper::IPV6_ADDRESS_LENGTH || $cidr < 0) {
                     return [$this->wrongCidr, []];
                 }
             } else {
                 $isCidrDefault = true;
-                $cidr = static::IPV6_ADDRESS_LENGTH;
+                $cidr = IpHelper::IPV6_ADDRESS_LENGTH;
             }
 
             if (!$this->validateIPv6($ip)) {
@@ -376,12 +377,12 @@ class IpValidator extends Validator
             }
         } else {
             if ($cidr !== null) {
-                if ($cidr > static::IPV4_ADDRESS_LENGTH || $cidr < 0) {
+                if ($cidr > IpHelper::IPV4_ADDRESS_LENGTH || $cidr < 0) {
                     return [$this->wrongCidr, []];
                 }
             } else {
                 $isCidrDefault = true;
-                $cidr = static::IPV4_ADDRESS_LENGTH;
+                $cidr = IpHelper::IPV4_ADDRESS_LENGTH;
             }
             if (!$this->validateIPv4($ip)) {
                 return [$this->message, []];
@@ -412,10 +413,9 @@ class IpValidator extends Validator
      * @param string $ip the original IPv6
      * @return string the expanded IPv6
      */
-    public static function expandIPv6($ip)
+    private function expandIPv6($ip)
     {
-        $hex = unpack('H*hex', inet_pton($ip));
-        return substr(preg_replace('/([a-f0-9]{4})/i', '$1:', $hex['hex']), 0, -1);
+        return IpHelper::expandIPv6($ip);
     }
 
     /**
@@ -513,9 +513,9 @@ class IpValidator extends Validator
      * @param string $ip
      * @return int
      */
-    private static function getIpVersion($ip)
+    private function getIpVersion($ip)
     {
-        return strpos($ip, ':') === false ? 4 : 6;
+        return IpHelper::getIpVersion($ip);
     }
 
     /**
@@ -535,49 +535,9 @@ class IpValidator extends Validator
      * @param string $range subnet in CIDR format e.g. `10.0.0.0/8` or `2001:af::/64`
      * @return bool
      */
-    public static function inRange($ip, $cidr, $range)
+    private function inRange($ip, $cidr, $range)
     {
-        $ipVersion = static::getIpVersion($ip);
-        $binIp = static::ip2bin($ip);
-
-        $parts = explode('/', $range);
-        $net = array_shift($parts);
-        $range_cidr = array_shift($parts);
-
-
-        $netVersion = static::getIpVersion($net);
-        if ($ipVersion !== $netVersion) {
-            return false;
-        }
-        if ($range_cidr === null) {
-            $range_cidr = $netVersion === 4 ? static::IPV4_ADDRESS_LENGTH : static::IPV6_ADDRESS_LENGTH;
-        }
-
-        $binNet = static::ip2bin($net);
-        return substr($binIp, 0, $range_cidr) === substr($binNet, 0, $range_cidr) && $cidr >= $range_cidr;
-    }
-
-    /**
-     * Converts IP address to bits representation.
-     *
-     * @param string $ip
-     * @return string bits as a string
-     */
-    private static function ip2bin($ip)
-    {
-        if (static::getIpVersion($ip) === 4) {
-            return str_pad(base_convert(ip2long($ip), 10, 2), static::IPV4_ADDRESS_LENGTH, '0', STR_PAD_LEFT);
-        }
-
-        $unpack = unpack('A16', inet_pton($ip));
-        $binStr = array_shift($unpack);
-        $bytes = static::IPV6_ADDRESS_LENGTH / 8; // 128 bit / 8 = 16 bytes
-        $result = '';
-        while ($bytes-- > 0) {
-            $result = sprintf('%08b', isset($binStr[$bytes]) ? ord($binStr[$bytes]) : '0') . $result;
-        }
-
-        return $result;
+        return IpHelper::inRange($ip, $cidr, $range);
     }
 
     /**